chore : Rename email_verification module to token
