+++ dorandoran/users/views.py
@@ -5,7 +5,7 @@ from django.template.loader import render_to_string
 from django.utils.http import urlsafe_base64_encode
 from django.utils.html import strip_tags
 from django.utils.encoding import force_bytes
-from core.email_verification import account_activation_token
+from core.token import account_activation_token
 from .models import TeacherProfile, StudentProfile
 from .serializers import (
     TeacherProfileSerializer,
